requirements.txt
inquirer==3.1.2
mysql-connector-python==8.0.31
pandas==1.5.0
python-dotenv==0.21.0
requests==2.29.0
Faker==18.6.0